Fix Node.aux_properties inconsistent typing
aux_properties is supposed to be a Dict. However, a couple places in the code would initialize a Node by passing in a List.

I fixed this by moving some duplicate List->Dict code in to a method that is called whenever the aux_properties need to be updated, which is also used on Node __init__.