Correct malformed Climate commands
Bump Version to 1.1.2


Additional fixes to Climate Functions


Expand Event Results to include value, precision, and UOM

Expanded the Event Results to include more than just the control command.

For climate nodes, not all properties are stored in the actual node definition returned by the ISY. CLIHCS, CLIFS, UOM for example, are only seen as events, and the values are required to do anything useful with them.

A Class has been implemented to capture the other information, but still report back only a string so this is not a breaking change.